aider: Modified `contextMenuInteraction(_:configurationForMenuAtLocation:)` to return a `UIContextMenuConfiguration` that contains a `UIMenu`.
# Context:
USER: Cannot convert value of type 'UIContextMenuConfiguration?' to closure result type 'UIMenu'
ASSISTANT: It seems like you're trying to return a `UIMenu` from a closure that expects a `UIContextMenuConfiguration`. The `UIContextMenuConfiguration` is used to configure the context menu, and it can contain a `UIMenu` as one of its properties.

Here's how you can modify the `contextMenuInteraction(_:configurationForMenuAtLocation:)` method to return a `UIContextMenuConfiguration` that contains a `UIMenu`:

```swift
AiderTest/AiderTest/ChartViewController.swift
<<<<<<< ORIGINAL
    func contextMenuInteraction(_ interaction: UIContextMenuInteraction, configurationForMenuAtLocation location: CGPoint) -> UIContextMenuConfiguration? {
        return UIContextMenuConfiguration(identifier: nil, previewProvider: nil) { _ in
            let toggleAction = UIAction(title: "Toggle", image: nil, identifier: nil, discoverabilityTitle: nil, attributes: [], state: .off) { _ in
                self.toggleChartDisplay()
            }
            return UIMenu(title: "", image: nil, identifier: nil, options: [], children: [toggleAction])
        }
    }
=======
    func contextMenuInteraction(_ interaction: UIContextMenuInteraction, configurationForMenuAtLocation location: CGPoint) -> UIContextMenuConfiguration? {
        let toggleAction = UIAction(title: "Toggle", image: nil, identifier: nil, discoverabilityTitle: nil, attributes: [], state: .off) { _ in
            self.toggleChartDisplay()
        }
        let menu = UIMenu(title: "", image: nil, identifier: nil, options: [], children: [toggleAction])
        return UIContextMenuConfiguration(identifier: nil, previewProvider: nil, actionProvider: { _ in return menu })
    }
>>>>>>> UPDATED
```

In the updated code, I've created a `UIMenu` with the `toggleAction` and then returned a `UIContextMenuConfiguration` that uses this menu. The `actionProvider` closure of the `UIContextMenuConfiguration` is used to provide the actions (in this case, the `UIMenu`) when the context menu is displayed.