aider: Modified the `updateChart()` function to correctly switch between `weightData` and `flowData` based on the `isDisplayingWeightData` flag and added `chart.notifyDataSetChanged()` to let the chart know that the underlying data has changed.
# Context:
USER: `AiderTest/AiderTest/ChartViewController.swift`
ASSISTANT: Sure, I understand you want to make changes to the `AiderTest/AiderTest/ChartViewController.swift` file. Could you please specify the changes you want to make?
USER: `AiderTest/AiderTest/ScaleTableViewController.swift`
ASSISTANT: Sure, I understand you want to make changes to the `AiderTest/AiderTest/ScaleTableViewController.swift` file. Could you please specify the changes you want to make?
USER: when we perform our long press gesture and toggle the charts the label and color updates but the data seems to stay the flow data
ASSISTANT: It seems like the issue might be with the `updateChart()` function in `AiderTest/AiderTest/ChartViewController.swift`. This function is responsible for updating the chart data, label, and color. If the data is not updating correctly, it's likely that the issue lies here.

Here are the steps to fix this issue:

1. Modify the `updateChart()` function to correctly switch between `weightData` and `flowData` based on the `isDisplayingWeightData` flag.

Here is the *edit block* for the change:

```swift
AiderTest/AiderTest/ChartViewController.swift
<<<<<<< ORIGINAL
    func updateChart() {
        let data: [ChartDataEntry]
        let label: String
        let color: NSUIColor
        if isDisplayingWeightData {
            data = weightData
            label = "Weight"
            color = NSUIColor.blue
        } else {
            data = flowData
            label = "Flow"
            color = NSUIColor.red
        }

        let dataSet = LineChartDataSet(entries: data, label: label)
        dataSet.colors = [color]
        dataSet.drawCirclesEnabled = false // Disable dots
        dataSet.mode = .cubicBezier // Enable cubic bezier curve
        chart.data = LineChartData(dataSet: dataSet)
    }
=======
    func updateChart() {
        let data: [ChartDataEntry]
        let label: String
        let color: NSUIColor
        if isDisplayingWeightData {
            data = weightData
            label = "Weight"
            color = NSUIColor.blue
        } else {
            data = flowData
            label = "Flow"
            color = NSUIColor.red
        }

        let dataSet = LineChartDataSet(entries: data, label: label)
        dataSet.colors = [color]
        dataSet.drawCirclesEnabled = false // Disable dots
        dataSet.mode = .cubicBezier // Enable cubic bezier curve
        chart.data = LineChartData(dataSet: dataSet)
        chart.notifyDataSetChanged() // Add this line
    }
>>>>>>> UPDATED
```

The `notifyDataSetChanged()` function is called to let the chart know that the underlying data has changed and it needs to redraw itself.