aider: Registered a cell class for the identifier "Cell" in the table view.
# Context:
USER: "unable to dequeue a cell with identifier Cell - must register a nib or a class for the identifier or connect a prototype cell in a storyboard"
ASSISTANT: This error occurs because the table view is trying to dequeue a cell with the identifier "Cell", but it doesn't know what kind of cell that is. You need to register a cell class for that identifier.

Here are the changes:

```swift
AiderTest/AiderTest/ChartViewController.swift
<<<<<<< ORIGINAL
    func createTableView(at location: CGPoint) {
        let tableView = UITableView(frame: CGRect(x: location.x, y: location.y, width: 200, height: 300))
        tableView.delegate = self
        tableView.dataSource = self
        view.addSubview(tableView)
    }
=======
    func createTableView(at location: CGPoint) {
        let tableView = UITableView(frame: CGRect(x: location.x, y: location.y, width: 200, height: 300))
        tableView.delegate = self
        tableView.dataSource = self
        tableView.register(UITableViewCell.self, forCellReuseIdentifier: "Cell")
        view.addSubview(tableView)
    }
>>>>>>> UPDATED
```

Now, the table view knows that a cell with the identifier "Cell" is an instance of `UITableViewCell`.