aider: The tap gesture recognizer is now added to the chart view instead of the table view and is dismissed along with the menu.
# Context:
USER: I think that tap gesture recognizer only recognizez taps in the bounds of my table view because it was created inside the new table view. I still need it to only be created when after I do a long press gesture but it needs to be inside my chart view and it needs to be dismissed along with our little menu so it doesnt interfere with our long press gesture.
ASSISTANT: To make the tap gesture recognizer recognize taps in the bounds of the chart view, you can add it to the chart view instead of the table view. To dismiss it along with the menu, you can remove it from the chart view when the menu is dismissed.

Here are the changes:

```swift
AiderTest/AiderTest/ChartViewController.swift
<<<<<<< ORIGINAL
    func createTableView(at location: CGPoint) {
        let tableView = UITableView(frame: CGRect(x: location.x, y: location.y, width: 200, height: 300))
        tableView.delegate = self
        tableView.dataSource = self
        tableView.register(UITableViewCell.self, forCellReuseIdentifier: "Cell")
        view.addSubview(tableView)

        let tapRecognizer = UITapGestureRecognizer(target: self, action: #selector(handleTap))
        view.addGestureRecognizer(tapRecognizer)
    }

    @objc func handleTap(gesture: UITapGestureRecognizer) {
        for subview in view.subviews {
            if subview is UITableView {
                subview.removeFromSuperview()
            }
        }
    }
=======
    var tapRecognizer: UITapGestureRecognizer?

    func createTableView(at location: CGPoint) {
        let tableView = UITableView(frame: CGRect(x: location.x, y: location.y, width: 200, height: 300))
        tableView.delegate = self
        tableView.dataSource = self
        tableView.register(UITableViewCell.self, forCellReuseIdentifier: "Cell")
        view.addSubview(tableView)

        tapRecognizer = UITapGestureRecognizer(target: self, action: #selector(handleTap))
        chart.addGestureRecognizer(tapRecognizer!)
    }

    @objc func handleTap(gesture: UITapGestureRecognizer) {
        for subview in view.subviews {
            if subview is UITableView {
                subview.removeFromSuperview()
            }
        }
        if let recognizer = tapRecognizer {
            chart.removeGestureRecognizer(recognizer)
        }
    }
>>>>>>> UPDATED
```

Now, the tap gesture recognizer will recognize taps in the bounds of the chart view and will be dismissed along with the menu.