aider: Changed the height of the table view to be dynamically calculated based on the number of rows.
# Context:
USER: now, we need to dynamically change the size on our long press menu to only be as big as the number of rows it contains
ASSISTANT: To dynamically change the size of the table view based on the number of rows it contains, you can calculate the height of the table view based on the number of rows and the height of each row.

Here are the changes:

```swift
AiderTest/AiderTest/ChartViewController.swift
<<<<<<< ORIGINAL
    func createTableView(at location: CGPoint) {
        let tableView = UITableView(frame: CGRect(x: location.x, y: location.y, width: 200, height: 300))
        tableView.delegate = self
        tableView.dataSource = self
        tableView.register(UITableViewCell.self, forCellReuseIdentifier: "Cell")
        view.addSubview(tableView)

        tapRecognizer = UITapGestureRecognizer(target: self, action: #selector(handleTap))
        chart.addGestureRecognizer(tapRecognizer!)
    }
=======
    func createTableView(at location: CGPoint) {
        let rowHeight: CGFloat = 44 // Default row height
        let tableViewHeight = CGFloat(menuOptions.count) * rowHeight
        let tableView = UITableView(frame: CGRect(x: location.x, y: location.y, width: 200, height: tableViewHeight))
        tableView.delegate = self
        tableView.dataSource = self
        tableView.register(UITableViewCell.self, forCellReuseIdentifier: "Cell")
        view.addSubview(tableView)

        tapRecognizer = UITapGestureRecognizer(target: self, action: #selector(handleTap))
        chart.addGestureRecognizer(tapRecognizer!)
    }
>>>>>>> UPDATED
```

Now, the height of the table view will be equal to the number of rows times the height of each row. If you have a custom row height, replace `44` with your custom row height.