aider: Removed creation and addition of weight and flow labels to the view in the `setupUI` function.
# Context:
USER: Now, lets change up the flow of our app a little bit we can remove the labels from `AiderTest/AiderTest/ViewController.swift`
ASSISTANT: To remove the labels from `AiderTest/AiderTest/ViewController.swift`, we need to modify the `setupUI` function where the labels are created and added to the view.

Here is the change:

```swift
AiderTest/AiderTest/ViewController.swift
<<<<<<< ORIGINAL
    func setupUI() {
        // Create and setup the weight label
        weightLabel = UILabel()
        weightLabel.text = "weight: "
        weightLabel.translatesAutoresizingMaskIntoConstraints = false
        view.addSubview(weightLabel)

        // Create and setup the flow label
        flowLabel = UILabel()
        flowLabel.text = "flow: "
        flowLabel.translatesAutoresizingMaskIntoConstraints = false
        view.addSubview(flowLabel)
=======
    func setupUI() {
>>>>>>> UPDATED
```

This change will remove the creation and addition of the weight and flow labels to the view.