aider: Check if a scale is connected and navigate to `ScaleTableViewController` if not.
# Context:
USER: when we open our app we should automatically check if a scale is connected and if not navigate to `AiderTest/AiderTest/ScaleTableViewController.swift`
ASSISTANT: To implement this, we need to modify the `viewDidLoad` method in `AiderTest/AiderTest/ViewController.swift` to check if a scale is connected and navigate to `ScaleTableViewController` if not.

Here is the change:

```swift
AiderTest/AiderTest/ViewController.swift
<<<<<<< ORIGINAL
    override func viewDidLoad() {
        super.viewDidLoad()
        setupUI()
        NotificationCenter.default.addObserver(self, selector: #selector(onWeightUpdate(_:)), name: NSNotification.Name(rawValue: AcaiaScaleWeight), object: nil)
    }
=======
    override func viewDidLoad() {
        super.viewDidLoad()
        setupUI()
        NotificationCenter.default.addObserver(self, selector: #selector(onWeightUpdate(_:)), name: NSNotification.Name(rawValue: AcaiaScaleWeight), object: nil)

        // Check if a scale is connected
        if AcaiaManager.shared().connectedScale == nil {
            let scaleTableViewController = ScaleTableViewController()
            scaleTableViewController.delegate = self
            navigationController?.pushViewController(scaleTableViewController, animated: true)
        }
    }
>>>>>>> UPDATED
```

This change will automatically navigate to `ScaleTableViewController` if no scale is connected when the app is opened.