aider: Automatically connect to the scale if it's found during the scan.
# Context:
USER: can we revert back to the previous verstion of the code I'd like to change our approach
ASSISTANT: Sure, you can revert to a previous version of your code using git.

Here are the steps:

1. Use `git log` to find the commit hash of the version you want to revert to. The commit hash is a long string of numbers and letters.

2. Once you have the commit hash, use `git checkout <commit-hash>` to revert to that version of the code. Replace `<commit-hash>` with the actual commit hash.

3. If you want to create a new branch from this commit, you can use `git checkout -b <new-branch-name> <commit-hash>`. Replace `<new-branch-name>` with the name you want to give to the new branch.

Please note that any changes you made after the commit you're reverting to will be lost. If you want to keep these changes, you should create a new branch before reverting.
USER: I'd like it to find the scale name in the scan and if it finds a scale connect to it.
ASSISTANT: To automatically connect to a scale when it's found during the scan, we need to modify the `_didFinishScan` method in `AiderTest/AiderTest/ScaleTableViewController.swift`.

Here is the change:

```swift
AiderTest/AiderTest/ScaleTableViewController.swift
<<<<<<< ORIGINAL
    @objc private func _didFinishScan(notification: NSNotification) {
        tableView.refreshControl?.endRefreshing()
        tableView.reloadData()
    }
=======
    @objc private func _didFinishScan(notification: NSNotification) {
        tableView.refreshControl?.endRefreshing()
        tableView.reloadData()

        // Automatically connect to the scale if it's found
        if let scale = AcaiaManager.shared().scaleList.first(where: { $0.name == "YourScaleName" }) {
            scale.connect()
        }
    }
>>>>>>> UPDATED
```

This change will automatically connect to the scale if its name matches "YourScaleName" after the scan finishes. Please replace "YourScaleName" with the actual name of your scale.