aider: Automatically connect to the first scale found after the scan finishes.
# Context:
USER: I dont want to specify a scale name it should dynamically find scales in the peripheral and use the info from the scan to connect automatically. this way users dont need to know the name of thier scale it's all doen for them.
ASSISTANT: To automatically connect to a scale when it's found during the scan, we need to modify the `_didFinishScan` method in `AiderTest/AiderTest/ScaleTableViewController.swift`.

Here is the change:

```swift
AiderTest/AiderTest/ScaleTableViewController.swift
<<<<<<< ORIGINAL
    @objc private func _didFinishScan(notification: NSNotification) {
        tableView.refreshControl?.endRefreshing()
        tableView.reloadData()

        // Automatically connect to the scale if it's found
        if let scale = AcaiaManager.shared().scaleList.first(where: { $0.name == "YourScaleName" }) {
            scale.connect()
        }
    }
=======
    @objc private func _didFinishScan(notification: NSNotification) {
        tableView.refreshControl?.endRefreshing()
        tableView.reloadData()

        // Automatically connect to the first scale found
        if let scale = AcaiaManager.shared().scaleList.first {
            scale.connect()
        }
    }
>>>>>>> UPDATED
```

This change will automatically connect to the first scale found after the scan finishes.