aider: Navigate to `ChartViewController` instead of `ViewController` after connecting to a scale.
# Context:
USER: now. I'd like to add a pipeline for the user as they are brewing thier coffee. when we connected to a scale, instead of being taken to `ViewController` I would like to be taken to `chartViewController`
ASSISTANT: To navigate to `ChartViewController` instead of `ViewController` after connecting to a scale, we need to modify the `_didConnected` method in `AiderTest/AiderTest/ScaleTableViewController.swift`.

Here is the change:

```swift
AiderTest/AiderTest/ScaleTableViewController.swift
<<<<<<< ORIGINAL
    @objc private func _didConnected(notification: NSNotification) {
        _activityIndicatorView.stopAnimating()

        _timerForConnectTimeOut?.invalidate()
        _timerForConnectTimeOut = nil

        if let scale = AcaiaManager.shared().connectedScale {
            delegate?.scaleTableViewController(self, didSelect: scale)
        }

        navigationController?.popViewController(animated: true)
    }
=======
    @objc private func _didConnected(notification: NSNotification) {
        _activityIndicatorView.stopAnimating()

        _timerForConnectTimeOut?.invalidate()
        _timerForConnectTimeOut = nil

        if let scale = AcaiaManager.shared().connectedScale {
            delegate?.scaleTableViewController(self, didSelect: scale)
        }

        let chartViewController = ChartViewController()
        navigationController?.pushViewController(chartViewController, animated: true)
    }
>>>>>>> UPDATED
```

This change will navigate to `ChartViewController` instead of `ViewController` after a scale is connected.