aider: Check if a scale is connected and navigate to `ScaleTableViewController` if not, or start logging if a scale is connected when the start button is tapped.
# Context:
USER: now in my `chartViewController` when I press start it checks if I have a scale connected and if I dont it takes me to `scaleTableViewController` if I do it starts logiing
ASSISTANT: To implement this, we need to modify the `startButtonTapped` method in `AiderTest/AiderTest/ChartViewController.swift` to check if a scale is connected and navigate to `ScaleTableViewController` if not, or start logging if a scale is connected.

Here is the change:

```swift
AiderTest/AiderTest/ChartViewController.swift
<<<<<<< ORIGINAL
    @objc func startButtonTapped() {
        if isLogging {
            isLogging = false
            startButton.setTitle("Start", for: .normal)
            stopLogging()
        } else {
            isLogging = true
            startButton.setTitle("Stop", for: .normal)
            startLogging()
        }
    }
=======
    @objc func startButtonTapped() {
        if isLogging {
            isLogging = false
            startButton.setTitle("Start", for: .normal)
            stopLogging()
        } else {
            if AcaiaManager.shared().connectedScale == nil {
                let scaleTableViewController = ScaleTableViewController()
                navigationController?.pushViewController(scaleTableViewController, animated: true)
            } else {
                isLogging = true
                startButton.setTitle("Stop", for: .normal)
                startLogging()
            }
        }
    }
>>>>>>> UPDATED
```

This change will check if a scale is connected when the start button is tapped. If a scale is not connected, it will navigate to `ScaleTableViewController`. If a scale is connected, it will start logging.