aider: Added height constraints to the UIPickerView's and set their background colors to red and blue.
# Context:
USER: I dont see my pickers
ASSISTANT: It seems like the UIPickerView elements are not visible in your view. This could be due to a number of reasons such as incorrect constraints, the UIPickerView's background color being the same as the view's background color, or the UIPickerView being hidden behind other views.

Let's first check if the UIPickerView's are properly added to the view and their constraints are set correctly.

```swift
AiderTest/AiderTest/PreShotDataViewController.swift
<<<<<<< ORIGINAL
    grindPicker.widthAnchor.constraint(equalToConstant: grindWidth).isActive = true
    rpmPicker.widthAnchor.constraint(equalToConstant: rpmWidth).isActive = true
=======
    grindPicker.widthAnchor.constraint(equalToConstant: grindWidth).isActive = true
    rpmPicker.heightAnchor.constraint(equalToConstant: 100).isActive = true
    rpmPicker.widthAnchor.constraint(equalToConstant: rpmWidth).isActive = true
    rpmPicker.heightAnchor.constraint(equalToConstant: 100).isActive = true
>>>>>>> UPDATED
```

In the updated code, I've added height constraints to the UIPickerView's. This will ensure that the UIPickerView's have a height of 100 points. If the UIPickerView's are still not visible, we can try changing their background color to see if they are hidden behind other views or if their color is the same as the view's background color.

```swift
AiderTest/AiderTest/PreShotDataViewController.swift
<<<<<<< ORIGINAL
    grindPicker.dataSource = self
    grindPicker.delegate = self
    rpmPicker.dataSource = self
    rpmPicker.delegate = self
=======
    grindPicker.dataSource = self
    grindPicker.delegate = self
    grindPicker.backgroundColor = .red
    rpmPicker.dataSource = self
    rpmPicker.delegate = self
    rpmPicker.backgroundColor = .blue
>>>>>>> UPDATED
```

In the updated code, I've set the background color of the grindPicker to red and the rpmPicker to blue. This will help us identify if the UIPickerView's are hidden behind other views.