aider: Set `rpmPicker.isHidden` to `true` to temporarily hide the `rpmPicker` and verify if it's obstructing the `grindPicker`.
# Context:
USER: the only element on screen I can interact with is the rpm picker. could this be why I'm not seeing the`grindPicker` is the subview created by the `rpmPicker` obstructing everything else?
ASSISTANT: It's possible that the `rpmPicker` is obstructing the `grindPicker` if it's placed on top of it or if its size and position cover the `grindPicker`.

To verify this, we can temporarily change the `isHidden` property of the `rpmPicker` to `true` and see if the `grindPicker` becomes visible.

```swift
AiderTest/AiderTest/PreShotDataViewController.swift
<<<<<<< ORIGINAL
    rpmPicker.dataSource = self
    rpmPicker.delegate = self
    rpmPicker.backgroundColor = .blue
=======
    rpmPicker.isHidden = true
    rpmPicker.dataSource = self
    rpmPicker.delegate = self
    rpmPicker.backgroundColor = .blue
>>>>>>> UPDATED
```

In the updated code, I've set `rpmPicker.isHidden` to `true`. This will hide the `rpmPicker` and allow us to see if it's obstructing the `grindPicker`. If the `grindPicker` becomes visible after this change, then the `rpmPicker` was indeed obstructing it. In that case, you'll need to adjust the size and position of the `rpmPicker` and `grindPicker` so they don't overlap.