aider: Added dose variable update with current weight from the scale until set button is pressed and locked in, and updated dose label to reflect the value of the dose variable.
# Context:
USER: in my `PreShotDataViewController` can we have the dose variable update with the current weight we're recieving from the scale untill we press the set button which will lock in the dose variable to whatever value it was when set was pressed. also the dose label should reflect the value of the dose variable
ASSISTANT: To achieve this, we need to make a few changes:

1. Add a variable to keep track of whether the dose has been set.
2. Update the dose variable with the weight from the scale in a method that is called when new data is received from the scale.
3. Update the dose label whenever the dose variable changes.
4. When the set button is pressed, change the variable from step 1 to indicate that the dose has been set and stop updating the dose variable.

Here are the changes:

```swift
AiderTest/AiderTest/PreShotDataViewController.swift
<<<<<<< ORIGINAL
    var Dose: Double = 0.0
    var Grind: Double = 17.0
    var RPM: Double = 100.0
    var PreWet: Bool = false
=======
    var Dose: Double = 0.0 {
        didSet {
            doseLabel.text = "Dose: \(Dose)"
        }
    }
    var doseSet: Bool = false
    var Grind: Double = 17.0
    var RPM: Double = 100.0
    var PreWet: Bool = false
>>>>>>> UPDATED
```

In the updated code, I've added a `didSet` observer to the `Dose` variable that updates the `doseLabel` text whenever `Dose` changes. I've also added a `doseSet` variable to keep track of whether the dose has been set.

```swift
AiderTest/AiderTest/PreShotDataViewController.swift
<<<<<<< ORIGINAL
    let doseButton = UIButton()
=======
    let doseButton: UIButton = {
        let button = UIButton()
        button.addTarget(self, action: #selector(doseButtonPressed), for: .touchUpInside)
        return button
    }()
>>>>>>> UPDATED
```

In the updated code, I've added a target to the `doseButton` that calls a method `doseButtonPressed` when the button is pressed.

```swift
AiderTest/AiderTest/PreShotDataViewController.swift
<<<<<<< ORIGINAL
    // MARK: - UIPickerViewDelegate
=======
    @objc func doseButtonPressed() {
        doseSet = true
    }

    // This method should be called when new data is received from the scale
    func updateDose(with weight: Double) {
        if !doseSet {
            Dose = weight
        }
    }

    // MARK: - UIPickerViewDelegate
>>>>>>> UPDATED
```

In the updated code, I've added a `doseButtonPressed` method that sets `doseSet` to `true` when the dose button is pressed. I've also added an `updateDose(with:)` method that updates the `Dose` variable with the weight from the scale if `doseSet` is `false`. This method should be called when new data is received from the scale.